refactor(ai): split diplomatic_candidate_scoring for line-size gate (#2509)
Extract offer-peace and declare-war scoring into part files so the
library stays under the 1000 non-comment line repo-lint limit.
